requirements.txt
django-simple-history

djangorestframework
